move versions back to avoid circular deps
