+++ pkg/test/framework/resource/settings.go
@@ -19,8 +19,6 @@ import (
 	"path"
 	"strings"
 
-	"istio.io/istio/pkg/test/framework/components/istio"
-
 	"github.com/google/uuid"
 
 	"istio.io/istio/pkg/test/framework/label"
@@ -85,7 +83,7 @@ type Settings struct {
 
 	// Versions are the Istio versions accessible to each of the clusters.
 	// represented as comma-separated versions, such as "1.7.3,1.8.2,1.9.0"
-	Versions istio.Versions
+	Versions Versions
 }
 
 // RunDir is the name of the dir to output, for this particular run.
