+++ pkg/test/framework/resource/version.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package istio
+package resource
 
 import (
 	"strings"
